+++ MapboxVision/VisionManager/VisionReplayManager.swift
@@ -16,6 +16,20 @@ import Foundation
  Do NOT use session replay in production application.
  */
 public final class VisionReplayManager: BaseVisionManager {
+    // MARK: - Public
+
+    /// Delegate for `VisionManager`. Delegate is held as a weak reference.
+    public weak var delegate: VisionManagerDelegate? {
+        get {
+            return baseDelegate
+        }
+        set {
+            baseDelegate = newValue
+        }
+    }
+
+    // MARK: Lifetime
+
     /**
      Fabric method for creating a `VisionReplayManager` instance.
 
@@ -45,9 +59,23 @@ public final class VisionReplayManager: BaseVisionManager {
 
      - Important: Do NOT call this method more than once or after `destroy` is called.
 
-     - Parameter delegate: Delegate for `VisionReplayManager`. Delegate is held as a strong reference until `stop` is called.
+     - Parameter delegate: Delegate for `VisionReplayManager`.
+         Until MapboxVision 0.9.0 delegate was held as a strong reference until `stop` is called.
+         Since MapboxVision 0.9.0 delegate is held as a weak reference and is not reset on `stop`.
      */
-    public func start(delegate: VisionManagerDelegate? = nil) {
+    @available(*, deprecated, message: "This will be removed in 0.10.0. Use method start() instead and set delegate as property.")
+    public func start(delegate: VisionManagerDelegate?) {
+        baseDelegate = delegate
+        start()
+    }
+
+    /**
+     Start delivering events from `VisionReplayManager`.
+     Calling `start` on already started or destroyed instance is considered a mistake.
+
+     - Important: Do NOT call this method more than once or after `destroy` is called.
+     */
+    public func start() {
         switch state {
         case .uninitialized:
             assertionFailure("VisionManager should be initialized before starting")
@@ -56,7 +84,6 @@ public final class VisionReplayManager: BaseVisionManager {
             assertionFailure("VisionManager is already started")
             return
         case .initialized, .stopped:
-            self.delegate = delegate
             state = .started
         }
 
@@ -77,7 +104,6 @@ public final class VisionReplayManager: BaseVisionManager {
         pause()
 
         state = .stopped
-        delegate = nil
     }
 
     /**
@@ -105,6 +131,7 @@ public final class VisionReplayManager: BaseVisionManager {
 
         dependencies.player.add(observer: self)
         dependencies.player.delegate = self
+        dependencies.native.videoSource = VideoSourceObserverProxy(withVideoSource: videoSource)
 
         state = .initialized
     }
@@ -146,7 +173,7 @@ public final class VisionReplayManager: BaseVisionManager {
     }
 
     private func resume() {
-        dependencies.native.start(self)
+        dependencies.native.start()
         dependencies.player.start()
     }